fix: symlink files (#5)

+++ scripts/symlink.sh
@@ -10,43 +10,45 @@ root_dir=$(dirname "$scripts_dir")
 
 # Function to backup existing file or directory and create a symlink
 backup_and_link() {
-	local source=""
-	local target=""
+    local source=""
+    local target=""
 
-	# Parse options
-	while [[ $# -gt 0 ]]; do
-		case "$1" in
-		--source)
-			source="$2"
-			shift 2
-			;;
-		--target)
-			target="$2"
-			shift 2
-			;;
-		*)
-			echo "Usage: backup_and_link --source <source_path> --target <target_path>"
-			return 1
-			;;
-		esac
-	done
+    # Parse options
+    while [[ $# -gt 0 ]]; do
+        case "$1" in
+        --source)
+            source="$2"
+            shift 2
+            ;;
+        --target)
+            target="$2"
+            shift 2
+            ;;
+        *)
+            echo "Usage: backup_and_link --source <source_path> --target <target_path>"
+            return 1
+            ;;
+        esac
+    done
 
-	local target_basename=
-	target_basename=$(basename "$target")
+    local target_basename=
+    target_basename=$(basename "$target")
 
-	# Check if the target exists and create a backup
-	if [ -e "$target" ] && [ ! -L "$target" ]; then
-		# shellcheck disable=SC2155
-		local backup="$root_dir/backup/${target_basename}_$(date +%Y%m%d%H%M%S)"
-		mv "$target" "$backup"
-		echo "    Backup of '$target' created at '$backup'"
-	fi
+    # Check if the target exists and create a backup
+    if [ -e "$target" ] && [ ! -L "$target" ]; then
+        # shellcheck disable=SC2155
+        local backup="$root_dir/backup/${target_basename}_$(date +%Y%m%d%H%M%S)"
+        mv "$target" "$backup"
+        echo "    Backup of '$target' created at '$backup'"
+    fi
 
-	# Force-create the symlink
-	ln -sf "$source" "$target"
-	# Prevent recursion
-	rm -f "$source/$target_basename"
-	echo "    Symlink created: '$target' -> '$source'"
+    # Force-create the symlink
+    ln -sf "$source" "$target"
+    # Prevent recursion
+    if [[ -d "$source" && "$source/$target_basename" == "$target" ]]; then
+        rm -f "$source/$target_basename"
+    fi
+    echo "    Symlink created: '$target' -> '$source'"
 }
 
 # -- Link .config items
@@ -54,8 +56,8 @@ backup_and_link() {
 config_files=$(ls -A "$root_dir/config")
 
 for config in $config_files; do
-	echo "[INFO] Symlink config: $config"
-	backup_and_link --source "$root_dir/config/$config" --target "$HOME/.config/$config"
+    echo "[INFO] Symlink config: $config"
+    backup_and_link --source "$root_dir/config/$config" --target "$HOME/.config/$config"
 done
 
 # -- Link dot-home-thingies
@@ -63,6 +65,6 @@ done
 dot_home_files=$(ls -A "$root_dir/dot-home-thingies")
 
 for config in $dot_home_files; do
-	echo "[INFO] Symlink config: $config"
-	backup_and_link --source "$root_dir/dot-home-thingies/$config" --target "$HOME/$config"
+    echo "[INFO] Symlink config: $config"
+    backup_and_link --source "$root_dir/dot-home-thingies/$config" --target "$HOME/$config"
 done
